Update schema-deletion-guidelines.rst
@joel-hamill

Here is my edit--wasn't sure about the "1" value in lines 12, 16 and 20--not sure why they are carried over to the following lines. Thought it could be typo, but wasn't sure about the command syntax, so I edited them for consistency, so they might be wrong, but they appear consistently wrong. ;-)
Joyce Fee